feat(app-admin): render ProblemDetail description as safe-markdown via shared web-kit Markdown
Extract `renderMarkdownToSafeHtml` (marked + DOMPurify allowlist, #661/#865) from
participant-portal into `@tenkacloud/web-kit` and add a `<Markdown>` component that
renders the sanitized HTML through a single `dangerouslySetInnerHTML` choke point.
admin-console ProblemDetail now renders `description` as markdown (headings / code
blocks / `![](https://…)` images) instead of plain `pre-wrap`; participant-portal
re-exports from web-kit (DRY, behavior unchanged).

Closes #1700
