+++ apps/participant-portal/package.json
@@ -18,9 +18,7 @@
     "@tenkacloud/format": "workspace:*",
     "@tenkacloud/web-kit": "workspace:*",
     "@tenkacloud/portal-plugin-sdk": "workspace:*",
-    "dompurify": "^3.4.4",
     "http-status-codes": "^2.3.0",
-    "marked": "^18.0.3",
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
     "react-router": "^7.15.1",
