+++ packages/web-kit/src/index.ts
@@ -26,6 +26,7 @@ export {
   resolveKey,
 } from "./i18n";
 export { LoadingState, type LoadingStateProps } from "./LoadingState";
+export { Markdown, type MarkdownProps, renderMarkdownToSafeHtml } from "./markdown";
 export { StatusBadge, type StatusBadgeProps, type StatusTone, statusToTone } from "./StatusBadge";
 export { useNowMs } from "./useNowMs";
 export { type UsePollingOptions, usePolling } from "./usePolling";
